@@ -1,6 +1,7 @@
 """Database infrastructure."""
 
-from .base import Base, TimestampMixin
+from shared.database.base import Base, TimestampMixin
+
 from .session import get_database_url, get_db, get_engine, get_session_factory, init_db
 
 __all__ = [
