@@ -1,16 +1,11 @@
-"""Blob storage module."""
+"""Blob storage module - server factory wrapping shared implementations."""
 
-from .blob_storage import BlobStorage
+from shared.storage import BlobStorage
+
 from .blob_storage_factory import BlobStorageType, create_blob_storage
-from .file_system_blob_storage import FileSystemBlobStorage
-from .in_memory_blob_storage import InMemoryBlobStorage
-from .minio_blob_storage import MinioBlobStorage
 
 __all__ = [
     "BlobStorage",
-    "FileSystemBlobStorage",
-    "InMemoryBlobStorage",
-    "MinioBlobStorage",
     "BlobStorageType",
     "create_blob_storage",
 ]
