@@ -6,13 +6,11 @@ from unittest.mock import patch
 
 import pytest
 
-from src.infrastructure.storage import (
-    BlobStorageType,
-    FileSystemBlobStorage,
-    InMemoryBlobStorage,
-    MinioBlobStorage,
-    create_blob_storage,
-)
+from shared.storage import BlobStorage
+from shared.storage.file_system_blob_storage import FileSystemBlobStorage
+from shared.storage.in_memory_blob_storage import InMemoryBlobStorage
+from shared.storage.minio_storage import MinIOBlobStorage
+from src.infrastructure.storage import BlobStorageType, create_blob_storage
 
 
 class TestBlobStorageFactory:
@@ -21,14 +19,14 @@ class TestBlobStorageFactory:
     def test_create_s3_storage_explicitly(self) -> None:
         """Test creating S3/MinIO storage with explicit type."""
         with (
-            patch("src.config.S3_ENDPOINT_URL", "http://minio:9000"),
+            patch("src.config.S3_ENDPOINT_URL", "minio:9000"),
             patch("src.config.S3_ACCESS_KEY", "test_key"),
             patch("src.config.S3_SECRET_KEY", "test_secret"),
             patch("src.config.S3_BUCKET_NAME", "test-bucket"),
-            patch("src.infrastructure.storage.minio_blob_storage.boto3.client"),
+            patch("shared.storage.minio_storage.Minio"),
         ):
             storage = create_blob_storage(storage_type=BlobStorageType.S3)
-            assert isinstance(storage, MinioBlobStorage)
+            assert isinstance(storage, MinIOBlobStorage)
 
     def test_create_file_system_storage_explicitly(self) -> None:
         """Test creating file system storage with explicit type."""
@@ -52,14 +50,14 @@ class TestBlobStorageFactory:
         """Test creating storage from config (S3 type)."""
         with (
             patch("src.config.BLOB_STORAGE_TYPE", "s3"),
-            patch("src.config.S3_ENDPOINT_URL", "http://localhost:9000"),
+            patch("src.config.S3_ENDPOINT_URL", "localhost:9000"),
             patch("src.config.S3_ACCESS_KEY", "admin"),
             patch("src.config.S3_SECRET_KEY", "password"),
             patch("src.config.S3_BUCKET_NAME", "files"),
-            patch("src.infrastructure.storage.minio_blob_storage.boto3.client"),
+            patch("shared.storage.minio_storage.Minio"),
         ):
             storage = create_blob_storage()
-            assert isinstance(storage, MinioBlobStorage)
+            assert isinstance(storage, MinIOBlobStorage)
 
     def test_create_storage_from_config_file_system(self) -> None:
         """Test creating storage from config (file system type)."""
