@@ -3,25 +3,24 @@
 from unittest.mock import MagicMock, patch
 
 import pytest
-
-from src.infrastructure.llm.claude_provider import ClaudeLlmProvider
-from src.infrastructure.llm.factory import create_llm_provider, get_available_providers
-from src.infrastructure.llm.huggingface_provider import HuggingFaceLlmProvider
-from src.infrastructure.llm.ollama import OllamaLlmProvider
-from src.infrastructure.llm.openai_provider import OpenAiLlmProvider
+from shared.llm.claude_provider import ClaudeLlmProvider
+from shared.llm.factory import create_llm_provider, get_available_providers
+from shared.llm.huggingface_provider import HuggingFaceLlmProvider
+from shared.llm.ollama import OllamaLlmProvider
+from shared.llm.openai_provider import OpenAiLlmProvider
 
 
 class TestCreateLlmProvider:
     """Tests for create_llm_provider factory function."""
 
-    @patch("src.infrastructure.llm.factory.OllamaLlmProvider")
+    @patch("shared.llm.factory.OllamaLlmProvider")
     def test_create_ollama_provider(self, mock_ollama: MagicMock) -> None:
         """Test creating Ollama provider."""
         provider = create_llm_provider("ollama")
 
         assert isinstance(provider, type(mock_ollama.return_value))
 
-    @patch("src.infrastructure.llm.factory.OllamaLlmProvider")
+    @patch("shared.llm.factory.OllamaLlmProvider")
     def test_create_ollama_with_custom_params(self, mock_ollama: MagicMock) -> None:
         """Test creating Ollama with custom parameters."""
         create_llm_provider("ollama", base_url="http://custom:11434", model_name="custom-model")
@@ -30,42 +29,42 @@ class TestCreateLlmProvider:
             base_url="http://custom:11434", model_name="custom-model"
         )
 
-    @patch("src.infrastructure.llm.factory.OpenAiLlmProvider")
+    @patch("shared.llm.factory.OpenAiLlmProvider")
     def test_create_openai_provider(self, mock_openai: MagicMock) -> None:
         """Test creating OpenAI provider."""
         provider = create_llm_provider("openai")
 
         assert isinstance(provider, type(mock_openai.return_value))
 
-    @patch("src.infrastructure.llm.factory.OpenAiLlmProvider")
+    @patch("shared.llm.factory.OpenAiLlmProvider")
     def test_create_openai_with_custom_params(self, mock_openai: MagicMock) -> None:
         """Test creating OpenAI with custom parameters."""
         create_llm_provider("openai", api_key="custom-key", model_name="gpt-4")
 
         mock_openai.assert_called_once_with(api_key="custom-key", model_name="gpt-4")
 
-    @patch("src.infrastructure.llm.factory.ClaudeLlmProvider")
+    @patch("shared.llm.factory.ClaudeLlmProvider")
     def test_create_claude_provider(self, mock_claude: MagicMock) -> None:
         """Test creating Claude provider."""
         provider = create_llm_provider("claude")
 
         assert isinstance(provider, type(mock_claude.return_value))
 
-    @patch("src.infrastructure.llm.factory.ClaudeLlmProvider")
+    @patch("shared.llm.factory.ClaudeLlmProvider")
     def test_create_claude_with_custom_params(self, mock_claude: MagicMock) -> None:
         """Test creating Claude with custom parameters."""
         create_llm_provider("claude", api_key="custom-key", model_name="claude-3-opus")
 
         mock_claude.assert_called_once_with(api_key="custom-key", model_name="claude-3-opus")
 
-    @patch("src.infrastructure.llm.factory.HuggingFaceLlmProvider")
+    @patch("shared.llm.factory.HuggingFaceLlmProvider")
     def test_create_huggingface_provider(self, mock_hf: MagicMock) -> None:
         """Test creating HuggingFace provider."""
         provider = create_llm_provider("huggingface")
 
         assert isinstance(provider, type(mock_hf.return_value))
 
-    @patch("src.infrastructure.llm.factory.HuggingFaceLlmProvider")
+    @patch("shared.llm.factory.HuggingFaceLlmProvider")
     def test_create_huggingface_with_custom_params(self, mock_hf: MagicMock) -> None:
         """Test creating HuggingFace with custom parameters."""
         create_llm_provider(
@@ -84,8 +83,8 @@ class TestCreateLlmProvider:
         with pytest.raises(ValueError, match="Unknown LLM provider: unknown"):
             create_llm_provider("unknown")
 
-    @patch("src.infrastructure.llm.factory.LLM_PROVIDER", "openai")
-    @patch("src.infrastructure.llm.factory.OpenAiLlmProvider")
+    @patch("shared.llm.factory.LLM_PROVIDER", "openai")
+    @patch("shared.llm.factory.OpenAiLlmProvider")
     def test_create_provider_uses_default(self, mock_openai: MagicMock) -> None:
         """Test that None provider uses default from config."""
         create_llm_provider(None)
@@ -96,7 +95,7 @@ class TestCreateLlmProvider:
 class TestGetAvailableProviders:
     """Tests for get_available_providers function."""
 
-    @patch("src.infrastructure.llm.factory.create_llm_provider")
+    @patch("shared.llm.factory.create_llm_provider")
     def test_get_available_providers_all_healthy(self, mock_create: MagicMock) -> None:
         """Test getting available providers when all are healthy."""
         mock_provider = type("MockProvider", (), {})()
@@ -113,7 +112,7 @@ class TestGetAvailableProviders:
         assert "claude" in result
         assert "huggingface" in result
 
-    @patch("src.infrastructure.llm.factory.create_llm_provider")
+    @patch("shared.llm.factory.create_llm_provider")
     def test_get_available_providers_with_errors(self, mock_create: MagicMock) -> None:
         """Test getting available providers when health check raises errors."""
 
@@ -148,8 +147,8 @@ class TestProviderIntegration:
     def test_create_all_providers_without_keys(self) -> None:
         """Test that all providers can be created without API keys."""
         with (
-            patch("src.infrastructure.llm.openai_provider.OpenAI"),
-            patch("src.infrastructure.llm.claude_provider.ANTHROPIC_AVAILABLE", True),
+            patch("shared.llm.openai_provider.OpenAI"),
+            patch("shared.llm.claude_provider.ANTHROPIC_AVAILABLE", True),
         ):
             ollama = create_llm_provider("ollama")
             openai = create_llm_provider("openai", api_key="")
@@ -164,8 +163,8 @@ class TestProviderIntegration:
     def test_provider_returns_correct_model_name(self) -> None:
         """Test that created providers return correct model names."""
         with (
-            patch("src.infrastructure.llm.openai_provider.OpenAI"),
-            patch("src.infrastructure.llm.claude_provider.ANTHROPIC_AVAILABLE", True),
+            patch("shared.llm.openai_provider.OpenAI"),
+            patch("shared.llm.claude_provider.ANTHROPIC_AVAILABLE", True),
         ):
             ollama = create_llm_provider("ollama", model_name="llama3.2")
             openai = create_llm_provider("openai", api_key="", model_name="gpt-4")
