@@ -1,10 +1,10 @@
 """Unit tests for UserService."""
 
 import pytest
-
-from src.domains.users.exceptions import UserAlreadyExistsError, UserAuthenticationError
 from shared.models import User
 from shared.repositories import UserRepository
+
+from src.domains.users.exceptions import UserAlreadyExistsError, UserAuthenticationError
 from src.domains.users.service import UserService
 
 
