@@ -3,5 +3,14 @@
 from shared.models.chat import ChatMessage, ChatSession
 from shared.models.document import Document
 from shared.models.user import User
+from shared.models.workspace import DefaultRagConfig, RagConfig, Workspace
 
-__all__ = ["User", "Document", "ChatSession", "ChatMessage"]
+__all__ = [
+    "User",
+    "Document",
+    "ChatSession",
+    "ChatMessage",
+    "Workspace",
+    "RagConfig",
+    "DefaultRagConfig",
+]
