@@ -2,5 +2,15 @@
 
 from shared.storage.interface import BlobStorage
 from shared.storage.minio_storage import MinIOBlobStorage
+from shared.storage.file_system_blob_storage import FileSystemBlobStorage
+from shared.storage.in_memory_blob_storage import InMemoryBlobStorage
+from shared.storage.factory import create_blob_storage, BlobStorageType
 
-__all__ = ["BlobStorage", "MinIOBlobStorage"]
+__all__ = [
+    "BlobStorage",
+    "MinIOBlobStorage",
+    "FileSystemBlobStorage",
+    "InMemoryBlobStorage",
+    "create_blob_storage",
+    "BlobStorageType",
+]
